+++ src/main/java/com/dh/TPIdiego_cata/service/IOdontologoService.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 public interface IOdontologoService {
     public Odontologo guardar (Odontologo odontologo);
 
-    public Optional<Odontologo> buscarPorId(Integer id);
+    public Optional<Odontologo> buscarPorId(Long id);
 
     public void actualizar(Odontologo odontologo);
 
-    public void eliminar(Integer id);
+    public void eliminar(Long id);
 
     public List<Odontologo> listarTodos();
 }
